Move most of main's code into a sponge pkg.
