--- app/sponge/main_test.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"testing"
-
-	"github.com/ardanlabs/kit/tests"
-)
-
-
-
